+++ example/go.work
@@ -1,0 +1,3 @@
+go 1.22.0
+
+use ..
