+++ example/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/IvanKorchmit/akevitt"
+	"github.com/IvanKorchmit/akevitt/plugins"
+)
+
+func main() {
+	akevitt.NewEngine().AddPlugin(plugins.DefaultPlugins()...)
+}
