+++ go.mod
@@ -2,30 +2,30 @@ module github.com/IvanKorchmit/akevitt
 
 require (
 	github.com/boltdb/bolt v1.3.1
-	github.com/gliderlabs/ssh v0.3.5
+	github.com/gliderlabs/ssh v0.3.7
 )
 
 require (
-	github.com/dlclark/regexp2 v1.4.0 // indirect
+	github.com/dlclark/regexp2 v1.11.0 // indirect
 	gitlab.com/tslocum/cbind v0.1.4 // indirect
-	golang.org/x/term v0.11.0 // indirect
+	golang.org/x/term v0.20.0 // indirect
 )
 
 require (
-	github.com/gdamore/encoding v1.0.0 // indirect
-	github.com/gdamore/tcell/v2 v2.6.0
+	github.com/gdamore/encoding v1.0.1 // indirect
+	github.com/gdamore/tcell/v2 v2.7.4
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
-	github.com/rivo/tview v0.0.0-20230814110005-ccc2c8119703
-	github.com/rivo/uniseg v0.4.4 // indirect
-	golang.org/x/text v0.12.0 // indirect
+	github.com/rivo/tview v0.0.0-20240524063012-037df494fb76
+	github.com/rivo/uniseg v0.4.7 // indirect
+	golang.org/x/text v0.15.0 // indirect
 )
 
 require (
 	github.com/anmitsu/go-shlex v0.0.0-20200514113438-38f4b401e2be // indirect
 	github.com/uaraven/logview v0.1.0
-	golang.org/x/crypto v0.12.0
-	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/crypto v0.23.0
+	golang.org/x/sys v0.20.0 // indirect
 )
 
-go 1.20
+go 1.22
