+++ stdUi.go
@@ -1,0 +1,28 @@
+package akevitt
+
+import (
+	"github.com/gdamore/tcell/v2"
+	"github.com/rivo/tview"
+	"github.com/uaraven/logview"
+)
+
+func AppendText(message string, chatlog *logview.LogView) {
+	ev := logview.NewLogEvent("message", message)
+	ev.Level = logview.LogLevelInfo
+	chatlog.AppendEvent(ev)
+	chatlog.SetFocusFunc(func() {
+		chatlog.Blur()
+	})
+	chatlog.ScrollToBottom()
+}
+
+func ErrorBox(message string, app *tview.Application, back tview.Primitive) {
+	result := tview.NewModal().SetText("Error!").SetText(message).SetTextColor(tcell.ColorRed).
+		SetBackgroundColor(tcell.ColorBlack).
+		AddButtons([]string{"Close"}).SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+		app.SetRoot(back, true)
+	})
+
+	result.SetBorder(true).SetBorderColor(tcell.ColorDarkRed)
+	app.SetRoot(result, true)
+}
